sorted choices in display and form snippets
It is now possible to display choices in alphabetical order in display and
form snippets.  This is handy for displaying codeset choices (which were
unordered in all display languages).  I will submit changes to scheming so that the same functionality is available for select, multiple_select and
multiple_checkbox snippets have the same functionality.