make ignore_empty play nice with repeating_text and last_release_date
I had applied the ignore_empty validator to repeating_text fields and last_release_date.
This didn't work because ignore_empty was executed before the repeating_text or
scheming_datetime validators and the field value was always "missing".

Changing the order in which the validators are applied fixed the problem.